fix: don't load snapshot in dev mode
